regroup pyfixture
